backend/requirements.txt
Flask==3.0.3
requests==2.32.3
azure-functions==1.21.3
pandas==2.2.3